shared: Fix definition of ipcrmall alias.
When the alias definition was being evaluated, the two spaces in
"s/  */ /g" were being merged (so the defined alias contained
"s/ */ /g" instead).  So instead of using two adjacent space character
literals in the string, use one space in parentheses followed by \1*.
That amounts to the same thing, but makes it through eval safely.